neorados/fifo: Rewrite as proper I/O object
Split nominal handle object and reference-counted
implementation. While we're at it, add lazy-open functionality.

(cherry picked from commit 3097297dd39432d172d69454419fa83a908075f6)
